add close method to FileHandle that returns its WriteMeasurements
